# src/main/java/com/finaro/dto/payment/BasePaymentResponseDto.java
package com.finaro.dto.payment;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.experimental.SuperBuilder;

@Data
@SuperBuilder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor
public class BasePaymentResponseDto {
    private Boolean approved;
}